halluceneate-backend/requirements.txt
fastapi
uvicorn[standard]
watchdog
PyMuPDF
python-docx
pytesseract
sentence-transformers
faiss-cpu
torch
python-dotenv